Conflicts:
	src/plugins/projectexplorer/toolchaintype.h
	src/plugins/qt4projectmanager/qt4projectmanagerconstants.h
	src/plugins/qt4projectmanager/qt4target.cpp
	src/plugins/qt4projectmanager/qtversionmanager.cpp

+++ src/plugins/projectexplorer/userfileaccessor.cpp
@@ -52,7 +52,10 @@
 using namespace ProjectExplorer;
 
 namespace {
+const char * const USERFILE_VERSION_KEY = "ProjectExplorer.Project.Updater.FileVersion";
 const char * const USERFILE_ENVIRONMENT_ID_KEY = "ProjectExplorer.Project.Updater.EnvironmentId";
+
+const char * const USERFILE_PROP = "qtcUserFileName";
 
 const char * const PROJECT_FILE_POSTFIX(".user");
 
@@ -378,20 +381,74 @@ QVariantMap UserFileAccessor::restoreSettings(Project *project)
     if (m_lastVersion < 0 || !project)
         return QVariantMap();
 
-    QString fileName = fileNameFor(project->file()->fileName());
-    if (!QFile::exists(fileName))
-        return QVariantMap();
-
     PersistentSettingsReader reader;
-    reader.load(fileName);
+
+    QString fileName = fileNameFor(project->file()->fileName());
+    if (!reader.load(fileName))
+        return QVariantMap();
 
     QVariantMap map(reader.restoreValues());
 
     // Get and verify file version:
-    const int fileVersion = map.value(QLatin1String(Constants::USERFILE_VERSION_KEY), 0).toInt();
-    if (fileVersion < m_firstVersion || fileVersion > m_lastVersion + 1) {
-        qWarning() << "File version" << fileVersion << "is not supported.";
+    int fileVersion = map.value(QLatin1String(USERFILE_VERSION_KEY), 0).toInt();
+    if (fileVersion < m_firstVersion) {
+        qWarning() << "File version" << fileVersion << "too old.";
         return QVariantMap();
+    }
+    bool skipBackup = false;
+    if (fileVersion > m_lastVersion + 1) {
+        int oldFileVersion = fileVersion;
+        int newestFileVersion = -1;
+        QString newestFileName;
+        QFileInfo qfi(fileName);
+        foreach (const QString &de, qfi.absoluteDir().entryList(
+                     QStringList() << (qfi.fileName() + ".*"))) {
+            QString extension = de.mid(qfi.fileName().length() + 1);
+            QString oldFileName = fileName + '.' + extension;
+            // This is a quick check to identify old versions.
+            for (fileVersion = m_lastVersion; fileVersion >= m_firstVersion; --fileVersion)
+                if (extension == m_handlers.value(fileVersion)->displayUserFileVersion()) {
+                    if (fileVersion > newestFileVersion) {
+                        newestFileVersion = fileVersion;
+                        newestFileName = oldFileName;
+                    }
+                    goto found;
+                }
+            // This is an expensive check needed to identify our own and newer versions
+            // (as we don't know what extensions will be assigned in the future).
+            if (reader.load(oldFileName)) {
+                map = reader.restoreValues();
+                fileVersion = map.value(QLatin1String(USERFILE_VERSION_KEY), 0).toInt();
+                if (fileVersion == m_lastVersion + 1) {
+                    fileName = oldFileName;
+                    goto gotFile;
+                }
+            }
+          found: ;
+        }
+        if (newestFileVersion < 0) {
+            qWarning() << "File version" << oldFileVersion << "too new.";
+            return QVariantMap();
+        }
+        fileName = newestFileName;
+        fileVersion = newestFileVersion;
+        if (!reader.load(fileName))
+            return QVariantMap();
+        map = reader.restoreValues();
+      gotFile:
+        QMessageBox::information(Core::ICore::instance()->mainWindow(),
+                QApplication::translate("ProjectExplorer::UserFileAccessor",
+                    "Using Old Project Settings File"),
+                QApplication::translate("ProjectExplorer::UserFileAccessor",
+                    "A versioned backup of the .user settings file will be used, "
+                    "because the non-versioned file was created by an incompatible "
+                    "newer version of Qt Creator.\n"
+                    "Project settings changes made since the last time this version "
+                    "of Qt Creator was used with this project are ignored, and changes "
+                    "made now will <b>not</b> be propagated to the newer version."),
+                QMessageBox::Ok);
+        project->setProperty(USERFILE_PROP, fileName);
+        skipBackup = true;
     }
 
     // Verify environment Id:
@@ -420,20 +477,15 @@ QVariantMap UserFileAccessor::restoreSettings(Project *project)
     }
 
     // Do we need to do a update?
-    if (fileVersion != m_lastVersion + 1) {
-        map.insert(QLatin1String(Constants::USERFILE_WAS_UPDATED_KEY), true);
+    if (fileVersion != m_lastVersion + 1 && !skipBackup) {
         const QString backupFileName = fileName + '.' + m_handlers.value(fileVersion)->displayUserFileVersion();
-        map.insert(QLatin1String(Constants::USERFILE_BACKUP_FILENAME_KEY), backupFileName);
         QFile::remove(backupFileName);  // Remove because copy doesn't overwrite
         QFile::copy(fileName, backupFileName);
     }
-    map.insert(QLatin1String(Constants::USERFILE_PREVIOUS_VERSION_KEY), fileVersion);
 
     // Update:
     for (int i = fileVersion; i <= m_lastVersion; ++i)
         map = m_handlers.value(i)->update(project, map);
-
-    map.insert(QLatin1String(Constants::USERFILE_VERSION_KEY), m_lastVersion + 1);
 
     return map;
 }
@@ -448,11 +500,13 @@ bool UserFileAccessor::saveSettings(Project *project, const QVariantMap &map)
     for (QVariantMap::const_iterator i = map.constBegin(); i != map.constEnd(); ++i)
         writer.saveValue(i.key(), i.value());
 
-    writer.saveValue(QLatin1String(Constants::USERFILE_VERSION_KEY), m_lastVersion + 1);
+    writer.saveValue(QLatin1String(USERFILE_VERSION_KEY), m_lastVersion + 1);
     writer.saveValue(QLatin1String(USERFILE_ENVIRONMENT_ID_KEY),
                      ProjectExplorerPlugin::instance()->projectExplorerSettings().environmentId.toString());
 
-    return writer.save(fileNameFor(project->file()->fileName()), "QtCreatorProject");
+    QString fileName = project->property(USERFILE_PROP).toString();
+    return writer.save(fileName.isEmpty() ? fileNameFor(project->file()->fileName()) : fileName,
+                       "QtCreatorProject");
 }
 
 void UserFileAccessor::addVersionHandler(UserFileVersionHandler *handler)
@@ -497,15 +551,17 @@ QVariantMap Version0Handler::convertBuildConfigurations(Project *project, const 
 
     // Find a valid Id to use:
     QString id;
-    if (project->id() == QLatin1String("GenericProjectManager.GenericProject"))
+    if (project->id() == QLatin1String("GenericProjectManager.GenericProject")) {
         id = QLatin1String("GenericProjectManager.GenericBuildConfiguration");
-    else if (project->id() == QLatin1String("CMakeProjectManager.CMakeProject"))
+    } else if (project->id() == QLatin1String("CMakeProjectManager.CMakeProject")) {
         id = QLatin1String("CMakeProjectManager.CMakeBuildConfiguration");
-    else if (project->id() == QLatin1String("Qt4ProjectManager.Qt4Project"))
+    } else if (project->id() == QLatin1String("Qt4ProjectManager.Qt4Project")) {
+        result.insert(QLatin1String("Qt4ProjectManager.Qt4BuildConfiguration.NeedsV0Update"), QVariant());
         id = QLatin1String("Qt4ProjectManager.Qt4BuildConfiguration");
-    else
+    } else {
         return QVariantMap(); // QmlProjects do not(/no longer) have BuildConfigurations,
                               // or we do not know how to handle this.
+    }
     result.insert(QLatin1String("ProjectExplorer.ProjectConfiguration.Id"), id);
 
     for (QVariantMap::const_iterator i = map.constBegin(); i != map.constEnd(); ++i) {
@@ -1651,9 +1707,9 @@ static QVariant version8EnvNodeTransform(const QVariant &var)
 {
     QString result = var.toString();
 
-    result.replace(QRegExp(QLatin1String("%SOURCEDIR%|\\$(SOURCEDIR\\b|{SOURCEDIR})")),
+    result.replace(QRegExp(QLatin1String("%SOURCEDIR%|\\$(SOURCEDIR\\b|\\{SOURCEDIR\\})")),
                    QLatin1String("%{sourceDir}"));
-    result.replace(QRegExp(QLatin1String("%BUILDDIR%|\\$(BUILDDIR\\b|{BUILDDIR})")),
+    result.replace(QRegExp(QLatin1String("%BUILDDIR%|\\$(BUILDDIR\\b|\\{BUILDDIR\\})")),
                    QLatin1String("%{buildDir}"));
 #ifdef Q_OS_UNIX
     for (int vStart = -1, i = 0; i < result.length(); ) {
